Question changes until finish

+++ app/src/main/java/com/example/moodtrackerproject/ui/mood/tests/StressTestState.kt
@@ -7,7 +7,7 @@ data class StressTestState(
     val question: QuestionBody = DataBaseRepository.listOfStressQs[0],
     val listOfOptions: List<OptionUiModel> = listOf(),
     val points: Int = 0,
-    val setQuestion: (Int) -> Unit = {},
+    val setQuestion: () -> Unit = {},
     val currQuestionNum: Int = 0,
     val moveQuestion: () -> Unit = {},
 )
